photoauth-backend/requirements.txt
fastapi
uvicorn
pillow
numpy
opencv-python
python-multipart
exifread
torch
torchvision
